457: Removed PopoverContainer component
